Adding tests to task feature

+++ apicultura/tests/v1/test_user.py
@@ -5,6 +5,7 @@ from apicultura.core.schemas import user_schema
 
 PATH = "/v1/users/"
 
+
 # TESTS TO CREATE ENDPOINT
 def test_successful_create_user(client):
     data = {
@@ -27,52 +28,48 @@ def test_successful_create_user(client):
 
 
 def test_unsuccessful400_creat_existing_username(client, user):
-    
-    data = {
-        'username': 'karl',
-        'email': user.email,
-        'password': 'secret'
-    }
-    response = client.post(
-        PATH,
-        json=data
-    )
+    data = {"username": "karl", "email": user["email"], "password": "secret"}
+    response = client.post(PATH, json=data)
 
     assert response.status_code == HTTPStatus.BAD_REQUEST
-    assert response.json() == {'detail': 'Username already exists on database'}
+    assert response.json() == {"detail": "Username already exists on database"}
+
 
 def test_unsuccessful400_creat_existing_email(client, user):
     data = {
-        'username': user.username,
-        'email': 'karl@pep.com',
-        'password': 'secret'
+        "username": user["username"],
+        "email": "karl@pep.com",
+        "password": "secret",
     }
-    response = client.post(
-        PATH,
-        json=data
-    )
+    response = client.post(PATH, json=data)
 
     assert response.status_code == HTTPStatus.BAD_REQUEST
-    assert response.json() == {'detail': 'Username already exists on database'}
+    assert response.json() == {"detail": "Username already exists on database"}
+
 
 # TESTS TO READ ENDPOINTS
-def test_read_user_successful(client, user):
-    
 
-    url = PATH + str(user.id)
+
+def test_read_user_successful(client, user):
+    url = PATH + str(user["id"])
 
     response = client.get(url)
 
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {
-        "username": user.username,
-        "email": user.email,
-        "id": user.id,
+        "username": user["username"],
+        "email": user["email"],
+        "id": user["id"],
     }
 
 
 def test_list_users_successful(client, users):
-    users_schema = {'users': [user_schema.UserOut.model_validate(user).model_dump() for user in users]}
+    users_schema = {
+        "users": [
+            user_schema.UserOut.model_validate(user).model_dump()
+            for user in users
+        ]
+    }
 
     url = PATH + "view/"
     response = client.get(url)
@@ -85,20 +82,18 @@ def test_list_users_successful(client, users):
 def test_update_user_successful(client, user, token):
     data = {"username": "Karl"}
 
-    url = PATH + str(user.id)
+    url = PATH + str(user["id"])
 
     response = client.put(
         url=url,
-        headers={
-            "Authorization": f"Bearer {token}"
-        }, 
+        headers={"Authorization": f"Bearer {token}"},
         json=data,
     )
     assert response.status_code == HTTPStatus.OK
     assert response.json() == {
         "username": data["username"],
-        "email": user.email,
-        "id": user.id,
+        "email": user["email"],
+        "id": user["id"],
     }
 
 
@@ -107,11 +102,8 @@ def test_update_user_unsuccessful_404(client, token):
     url = PATH + str(42)
 
     response = client.put(
-        url, 
-        headers = {
-            "Authorization": f"Bearer {token}"
-        },
-        json=data)
+        url, headers={"Authorization": f"Bearer {token}"}, json=data
+    )
 
     assert response.status_code == HTTPStatus.UNAUTHORIZED
     assert response.json() == {"detail": "Not enough permissions"}
@@ -119,9 +111,9 @@ def test_update_user_unsuccessful_404(client, token):
 
 def test_update_user_unseccessful_401(client, token, user_2):
     response = client.put(
-        PATH + str(user_2.id),
-        json = {"username": "update_test"},
-        headers = {"Authorization": f"Bearer {token}"}
+        PATH + str(user_2["id"]),
+        json={"username": "update_test"},
+        headers={"Authorization": f"Bearer {token}"},
     )
 
     assert response.status_code == HTTPStatus.UNAUTHORIZED
@@ -130,13 +122,9 @@ def test_update_user_unseccessful_401(client, token, user_2):
 
 # TESTS TO DELETE ENDPOINT
 def test_delete_user_successful(client, user, token):
-    url = PATH + str(user.id)
+    url = PATH + str(user["id"])
 
-    response = client.delete(
-        url,
-        headers={
-            "Authorization": f"Bearer {token}"
-        })
+    response = client.delete(url, headers={"Authorization": f"Bearer {token}"})
 
     assert response.status_code == HTTPStatus.NO_CONTENT
 
@@ -144,12 +132,7 @@ def test_delete_user_successful(client, user, token):
 def test_delete_user_unsuccessful_404(client, token):
     url = PATH + str(42)
 
-    response = client.delete(
-        url,
-        headers={
-            "Authorization":f"Bearer {token}"
-        }
-        )
+    response = client.delete(url, headers={"Authorization": f"Bearer {token}"})
 
     assert response.status_code == HTTPStatus.UNAUTHORIZED
     assert response.json() == {"detail": "Not enough permissions"}
